remove element provider from wrapping all app routes and pull it into a component that takes in the checkoutForm where its needed
